Create dataset.json
Turns out the actual file list can not be queried from the application server, pybis internally asks the datastore server directly. Therefore, there isn't a single json representation of the dataset and its files. In the openbismantic_api, I query the files separately and add it to the response object under 'files', we can change that if we want.